Maintain: Focus on fixing some small problems.
[string] Change parameter type.
    -- In order to support the normal use of bitmap under
        64 bits, correct the types of memcount and memdiff to long.

[Kconfig] Fix kernel compilation parameters.
    -- Modify the kernel optimization to o0, because x86 has some interrupt
        problems temporarily, which may be related to code optimization.

[io] Correct syntax errors.
    -- Fixed non generic write function for map pmio.

+++ include/string.h
@@ -55,8 +55,8 @@ extern int memcmp(const void *s1, const void *s2, size_t n);
  */
 extern int strcount(const char *str, char c);
 extern int strncount(const char *str, char c, size_t size);
-extern int memcount(const void *addr, int c, size_t size);
-extern void *memdiff(const void *addr, int c, size_t size);
+extern int memcount(const void *addr, long c, size_t size);
+extern void *memdiff(const void *addr, long c, size_t size);
 extern char *skip_spaces(const char *path);
 extern char *basename(const char *path);
